requirements.txt
lxml==4.6.3
python-dateutil==2.3
git+https://github.com/IATI/IATI-Rulesets.git@version-2.03#egg=iatirulesets
GitPython==3.1.14
